refactor: leaderboard typings
