+++ packages/client/src/pages/Leaderboard/LeaderboardField/SortMarker/typings.ts
@@ -1,4 +1,4 @@
-import { SortDirection } from '../../typings';
+import { SortDirection } from '../../../../store/features/leaderboard/typings';
 
 export type SortMarkerProps = {
   sortDirection: SortDirection;
