+++ packages/client/src/pages/Leaderboard/typings.ts
@@ -1,5 +1,3 @@
 export type LeaderboardProps = {
   header?: string;
 };
-export type SortOption = 'username' | 'score' | 'time' | 'matches';
-export type SortDirection = 'asc' | 'desc';
