+++ packages/client/src/store/features/leaderboard/typings.ts
@@ -1,9 +1,12 @@
-import { LeaderboardRowProps } from '../../../pages/Leaderboard/LeaderboardRow/typings';
-import { SortDirection, SortOption } from './../../../pages/Leaderboard/typings';
+import { LeaderboardRecord } from '../../../api/typings';
+
+export type SortOption = 'username' | 'score' | 'time' | 'matches';
+
+export type SortDirection = 'asc' | 'desc';
 
 export type LeaderboardState = {
   isLoading: boolean;
-  leaderboardTable: LeaderboardRowProps[];
+  leaderboardTable: Array<LeaderboardRecord>;
   sortOption: SortOption;
   sortDirection: SortDirection;
 };
